requirements.txt
pycryptodome
requests
bcrypt
pynput
